feat(contrib): ensure remote-access example compiles on Windows
